+++ backend/src/algorithm/utils.ts
@@ -2,6 +2,7 @@ import { Class, MatrixHashmap, Periods, RemainingLectures, Subject, Unavailable 
 import { hash, unhash } from './utils/hash';
 import { range, shuffleArray } from './utils/array';
 import { daysPerWeek } from './consts';
+import { validateSwap } from './utils/constraint';
 
 export function getTimesPerWeek({ timesPerWeek }: Subject) {
   return timesPerWeek instanceof Array ? timesPerWeek.reduce((a, b) => a + b, 0) : timesPerWeek;
@@ -18,44 +19,6 @@ export function getClassPeriodsPerDay(class_: Class) {
   });
 
   return shuffleArray(periodsPerDay);
-}
-
-function checkIfClassUnavailable(timetable: MatrixHashmap, class_: Class, dayIndex: number, periodIndex: number) {
-  return timetable[hash(class_.name)][dayIndex][periodIndex];
-}
-
-function checkIfTeacherUnavailable(unavailable: Unavailable, subject: Subject, dayIndex: number, periodIndex: number) {
-  return unavailable.teachers[hash(subject.teacher)][dayIndex][periodIndex];
-}
-
-function checkIfClassroomUnavailable(
-  unavailable: Unavailable,
-  subject: Subject,
-  dayIndex: number,
-  periodIndex: number,
-) {
-  return subject.classroom && unavailable.classrooms[hash(subject.classroom)][dayIndex][periodIndex];
-}
-
-function checkIfLectureQuantityFulfilled(remainingLectures: RemainingLectures, class_: Class, subject: Subject) {
-  return remainingLectures[class_.name][subject.name] === 0;
-}
-
-export function checkConstraints(
-  timetable: MatrixHashmap,
-  unavailable: Unavailable,
-  remainingLectures: RemainingLectures,
-  class_: Class,
-  subject: Subject,
-  dayIndex: number,
-  periodIndex: number,
-) {
-  return (
-    checkIfClassUnavailable(timetable, class_, dayIndex, periodIndex) || // class in not available
-    checkIfTeacherUnavailable(unavailable, subject, dayIndex, periodIndex) || // teacher is not available
-    checkIfClassroomUnavailable(unavailable, subject, dayIndex, periodIndex) || // classroom is not available
-    checkIfLectureQuantityFulfilled(remainingLectures, class_, subject) // all lectures for subject are in schedule
-  );
 }
 
 export function swapSlots(
@@ -77,16 +40,17 @@ export function swapSlots(
       const oldSubject = unhash<Subject>(timetable[hash(class_.name)][oldDayIndex][oldPeriodIndex] as string);
 
       if (
-        // validate old subject to new period
-        !(
-          checkIfClassUnavailable(timetable, class_, dayIndex, periodIndex) ||
-          checkIfTeacherUnavailable(unavailable, oldSubject, dayIndex, periodIndex) ||
-          checkIfClassroomUnavailable(unavailable, oldSubject, dayIndex, periodIndex)
-        ) &&
-        // validate new subject to old period
-        !(
-          checkIfTeacherUnavailable(unavailable, newSubject, oldDayIndex, oldPeriodIndex) ||
-          checkIfLectureQuantityFulfilled(remainingLectures, class_, newSubject)
+        validateSwap(
+          timetable,
+          unavailable,
+          remainingLectures,
+          class_,
+          newSubject,
+          dayIndex,
+          periodIndex,
+          oldSubject,
+          oldDayIndex,
+          oldPeriodIndex,
         )
       ) {
         // set new subject to old slot
